+++ @utahdts/utah-design-system/react/components/forms/TextArea.jsx
@@ -4,7 +4,6 @@ import { useRememberCursorPosition } from '../../hooks/useRememberCursorPosition
 import { joinClassNames } from '../../util/joinClassNames';
 import { IconButton } from '../buttons/IconButton';
 import { ErrorMessage } from './ErrorMessage';
-import { useFormContextInput } from './FormContext/useFormContextInput';
 import { RequiredStar } from './RequiredStar';
 
 /**
@@ -46,34 +45,22 @@ export function TextArea({
   wrapperClassName,
   ...rest
 }) {
-  const {
-    onChange: currentOnChange,
-    onClear: currentOnClear,
-    onFormKeyUp: currentOnFormKeyUp,
-    value: currentValue,
-  } = useFormContextInput({
-    defaultValue,
-    id,
-    onChange,
-    onClear,
-    value,
-  });
   const inputRef = /** @type {typeof useRef<HTMLTextAreaElement | null>} */ (useRef)(null);
 
   const onChangeSetCursorPosition = useRememberCursorPosition(inputRef, value || '');
 
   const { addPoliteMessage } = useAriaMessaging();
 
-  const showClearIcon = !!((isClearable || onClear) && currentValue);
+  const showClearIcon = !!((isClearable || onClear) && value);
 
   const clearInput = useCallback(
     /** @param {import('react').UIEvent} e */
     (e) => {
-      currentOnClear?.(e);
+      onClear?.(e);
       addPoliteMessage(`${label} input was cleared`);
       inputRef.current?.focus();
     },
-    [addPoliteMessage, currentOnClear, label]
+    [addPoliteMessage, onClear, label]
   );
 
   const checkKeyPressed = useCallback(
@@ -81,20 +68,18 @@ export function TextArea({
     (e) => {
       if (e.key === 'Escape' && showClearIcon) {
         clearInput(e);
-      } else {
-        currentOnFormKeyUp?.(e);
       }
     },
-    [clearInput, currentOnFormKeyUp, showClearIcon]
+    [clearInput, showClearIcon]
   );
 
   const onChangeCallback = useCallback(
     /** @param {import('react').ChangeEvent<HTMLElement>} e */
     (e) => {
       onChangeSetCursorPosition(e);
-      currentOnChange?.(e);
+      onChange?.(e);
     },
-    [onChangeSetCursorPosition, currentOnChange]
+    [onChangeSetCursorPosition, onChange]
   );
 
   return (
@@ -108,15 +93,16 @@ export function TextArea({
           aria-describedby={errorMessage ? `${id}-error` : undefined}
           aria-invalid={!!errorMessage}
           className={joinClassNames(className, showClearIcon ? 'text-area--clear-icon-visible' : null)}
+          defaultValue={defaultValue}
           disabled={isDisabled}
           id={id}
           name={name || id}
-          onChange={currentValue !== undefined ? onChangeCallback : undefined}
+          onChange={value !== undefined ? onChangeCallback : undefined}
           onKeyUp={checkKeyPressed}
           placeholder={placeholder ?? undefined}
           ref={inputRef}
           required={isRequired}
-          value={currentValue}
+          value={value}
           {...rest}
         />
         {